Support for bedrock converse and openai.
