+++ src/Agent/NewRelic/Agent/Extensions/NewRelic.Agent.Extensions/Llm/SupportabilityHelpers.cs
@@ -1,0 +1,63 @@
+// Copyright 2020 New Relic, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+using System;
+using System.Collections.Concurrent;
+using System.Text.RegularExpressions;
+using NewRelic.Agent.Api;
+
+namespace NewRelic.Agent.Extensions.Llm
+{
+    public static class SupportabilityHelpers
+    {
+        private const string OpenAiDateRemovalPattern = @"-\d{4}-\d{2}-\d{2}";
+
+        private static readonly ConcurrentDictionary<string, object> _seenModels = new();
+
+        public static void CreateModelIdSupportabilityMetrics(string model, IAgent agent)
+        {
+            // Only want to send this metric once-ish per model
+            if (!_seenModels.TryAdd(model, null))
+            {
+                return;
+            }
+
+            try
+            {
+                // Example standard bedrock: anthropic.claude-3-5-sonnet-20241022-v2:0
+                // Example infrerence bedrock: us.anthropic.claude-3-5-sonnet-20241022-v2:0
+                // Example alias openai: gpt-4o
+                // Example direct openai: gpt-4o-2024-11-20
+                var modelDetails = model.Split('.');
+                if (modelDetails.Length == 1) // OpenAI
+                {
+                    agent.RecordSupportabilityMetric("DotNet/LLM/openai/" + Regex.Replace(model, OpenAiDateRemovalPattern, string.Empty));
+                    return;
+                }
+
+                // Not OpenAI, checking for bedrock
+                if (modelDetails.Length != 2 && modelDetails.Length != 3)
+                {
+                    return;
+                }
+
+                // if there is a region, it will be the first part of the model id
+                var vendor = modelDetails.Length == 2 ? modelDetails[0] : modelDetails[1];
+
+                var modelIdDetails = modelDetails[1].Split(':')[0].Split('-');
+                if (modelIdDetails[0] == "nova" || modelIdDetails[0] == "titan" || modelIdDetails[0] == "claude") // first 2 - capture some extra details to narrow down support
+                {
+                    agent.RecordSupportabilityMetric("DotNet/LLM/" + vendor + "/" + modelIdDetails[0] + "-" + modelIdDetails[1]);
+                }
+                else // first only - any model that doesn't need the above extra details
+                {
+                    agent.RecordSupportabilityMetric("DotNet/LLM/" + vendor + "/" + modelIdDetails[0]);
+                }
+            }
+            catch (Exception ex) // if there is a problem, this will also only happen once-ish per model
+            {
+                agent.Logger.Finest($"Error creating model supportability metric for {model}: {ex.Message}");
+            }
+        }
+    }
+}
